backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.3
redis==3.3.11
sqlparse==0.3.0
Faker
django-cors-headers
django-extensions==3.1.0
django_filter==2.0.0
coverage==4.5.1
django_redis==4.9.0
gunicorn==19.6.0
